move entitlement to structs
